+++ js/web/components/ReleasePurchase.js
@@ -29,6 +29,7 @@ import rehypeSanitize from 'rehype-sanitize'
 import rehypeExternalLinks from 'rehype-external-links'
 import Royalty from './Royalty'
 import CreateGateModal from './CreateGateModal'
+import UnlockGateModal from './UnlockGateModal'
 
 const ReleasePurchase = (props) => {
   const { releasePubkey, metadata, router } = props
@@ -80,6 +81,7 @@ const ReleasePurchase = (props) => {
   const getGate = async () => {
     const { gates } = (await axios.get(`${process.env.NINA_API_ENDPOINT}/releases/${releasePubkey}/gates`)).data
     if (gates.length > 0) {
+      console.log('gates :>> ', gates);
       setGate(gates[0])
     }
   }
@@ -87,6 +89,7 @@ const ReleasePurchase = (props) => {
   useEffect(() => {
     getRelease(releasePubkey)
     getGate()
+    console.log('gate :>> ', gate);
     // const hubForRelease = async (releasePubkey) => {
     //   const result = await getPublishedHubForRelease(releasePubkey)
     //   setPublishedHub(result?.hub)
@@ -103,10 +106,6 @@ const ReleasePurchase = (props) => {
       setRelease(releaseState.tokenData[releasePubkey])
     }
   }, [releaseState.tokenData[releasePubkey]])
-
-  useEffect(() => {
-    getAmountHeld(releaseState.releaseMintMap[releasePubkey], releasePubkey)
-  }, [])
 
   useEffect(() => {
     setAmountHeld(collection[releasePubkey])
@@ -268,97 +267,6 @@ const ReleasePurchase = (props) => {
     setDownloadButtonString('Download')
   }
 
-  const handleUnlockGate = async () => {
-    try {
-      const message = new TextEncoder().encode(releasePubkey);
-      const messageBase64 = encodeBase64(message);
-      const signature = await wallet.signMessage(message);
-      const signatureBase64 = encodeBase64(signature);
-      const result = await axios.get(`${process.env.NINA_GATE_URL}/gate/${gate.id}?message=${encodeURIComponent(messageBase64)}&publicKey=${encodeURIComponent(wallet.publicKey.toBase58())}&signature=${encodeURIComponent(signatureBase64)}`)
-      
-      const response = await axios.get(result.data.url, {
-        method: "GET",
-        mode: "cors",
-        headers: {
-          "Content-Type": "application/octet-stream",
-        },
-        responseType: "blob",
-      });
-
-      if (response?.data) {
-        const a = document.createElement("a");
-        const url = window.URL.createObjectURL(response.data);
-        a.href = url;
-        a.download = release.gate.fileName;
-        a.click();
-      }  
-    } catch (error) { 
-      console.log('error: ', error)
-      // setResponse(error.response.data);
-    }
-  }
-
-  // const handleFileUpload = async () => {
-  //   console.log('process.env.NINA_GATE_URL :>> ', process.env.NINA_GATE_URL);
-  //   try {
-  //     const FILE_CHUNK_SIZE = 10_000_000
-
-  //     const message = new TextEncoder().encode(releasePubkey);
-  //     const messageBase64 = encodeBase64(message);
-  //     const signature = await wallet.signMessage(message);
-  //     const signatureBase64 = encodeBase64(signature);
-
-  //     const response = await axios.post(`${process.env.NINA_GATE_URL}/gate`, {
-  //       fileSize: file.size,
-  //       fileName: file.name,
-  //       publicKey: wallet.publicKey.toBase58(),
-  //       message: messageBase64,
-  //       signature: signatureBase64,
-  //       release: releasePubkey,
-  //     })
-  //     console.log('response: ', response.data)
-  //     const {
-  //       urls,
-  //       UploadId
-  //     } = response.data;
-  //     console.log('urls: ', urls)
-  //     const uploader = axios.create()
-  //     delete uploader.defaults.headers.put['Content-Type']
-
-  //     const keys = Object.keys(urls)
-  //     const promises = []
-    
-  //     for (const indexStr of keys) {
-  //       const index = parseInt(indexStr)
-  //       const start = index * FILE_CHUNK_SIZE
-  //       const end = (index + 1) * FILE_CHUNK_SIZE
-  //       const blob = index < keys.length
-  //         ? file.slice(start, end)
-  //         : file.slice(start)
-    
-  //       promises.push(axios.put(urls[index], blob))
-  //     }
-    
-  //     const resParts = await Promise.all(promises)
-  //     const result = resParts.map((part, index) => ({
-  //       ETag: part.headers.etag,
-  //       PartNumber: index + 1
-  //     }))
-  //     console.log('result: ', result)
-
-  //     const completeResponse = await axios.post(`${process.env.NINA_GATE_URL}/gate/finalize`, {
-  //       UploadId,
-  //       releasePublicKey: releasePubkey,
-  //       fileName: file.name,
-  //       fileSize: file.size,
-  //       parts: result
-  //     })
-  //     getGate()
-  //     console.log('completeResponse: ', completeResponse.data)
-  //   } catch (err) {
-  //     console.log(err)
-  //   }
-  // }
   return (
     <Box>
       <AmountRemaining variant="body2" align="left">
@@ -428,7 +336,7 @@ const ReleasePurchase = (props) => {
         </>
       )}
       {gate && (
-        <button onClick={() => handleUnlockGate()}>Unlock Gate</button>
+        <UnlockGateModal gate={gate} releasePubkey={releasePubkey} amountHeld={amountHeld} />
       )}
       {userIsRecipient && (
         <Royalty releasePubkey={releasePubkey} release={release} />
